+++ src/pages/product/components/ProductList.js
@@ -1,26 +1,13 @@
-import { getPublishedPosts } from "@/api/marketApi";
 import Loading from "@/components/Loading";
 import useModalStore from "@/utils/hooks/store/useModalStore";
-import { css } from "@emotion/react";
 import styled from "@emotion/styled";
+import { css } from "@emotion/react";
 import { useInfiniteQuery, useQueryClient } from "@tanstack/react-query";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 // 상품 리스트를 보여주는 공용 컴포넌트
-const ProductList = ({ param, queryKey }) => {
-  const [productList, setProductList] = useState([]);
-  const getProductList = async ({ pageParam = 1 }) => {
-    // pageParam : useInfiniteQuery의 getNextPageParam에서 반환해준 값 (=다음 불러올 페이지)
-    const resData = await getPublishedPosts(param);
-
-    const { page, lastPage, data: responseData } = resData.data;
-    setProductList((prevList) => [...prevList, ...responseData]);
-
-    // return은 아래 useInfiniteQuery에서 getNextPageParam으로 전달
-    // page 뜻을 전달하기 위해 이름 curPage로 전달
-    return { curPage: page, lastPage };
-  };
+const ProductList = ({ getProductList, productList, keyword, queryKey }) => {
   const queryClient = useQueryClient();
   const navigate = useNavigate();
   const [showTopButton, setShowTopButton] = useState(false); // "맨 위로 가기" 버튼의 표시 여부를 제어하는 상태
@@ -35,6 +22,7 @@ const ProductList = ({ param, queryKey }) => {
         // 마지막 페이지인 경우에는 더 이상 호출 불필요 , 마지막 페이지보다 전이면 +1 해준다
         // 여기서 return 하는 값은 pageParam으로 전달 됨
         return curPage < lastPage ? curPage + 1 : null;
+        ㅉ;
       }
     });
 
@@ -101,7 +89,7 @@ const ProductList = ({ param, queryKey }) => {
         window.removeEventListener("scroll", handleTop);
       };
     }
-  }, [param.keyword]);
+  }, [keyword]);
 
   useEffect(() => {
     return () => {
@@ -171,7 +159,7 @@ const ProductList = ({ param, queryKey }) => {
           <div>
             {!productList?.length && !isFetching && (
               <NoticeMsg>
-                {`${param.keyword ? "검색 결과가" : "상품이"} 없습니다.`}
+                {`${keyword ? "검색 결과가" : "상품이"} 없습니다.`}
               </NoticeMsg>
             )}
           </div>
